Fix all compilation errors when building for host

+++ include/Kyoto/Particles/IElement.hpp
@@ -14,8 +14,8 @@ public:
   virtual ~IElement(){};
 
   // -> CFrameDelayedKiller
-  void* operator new(unsigned long sz, const char*, const char*);
-  void* operator new(unsigned long sz) { return operator new(sz, "??(??)", nullptr); }
+  void* operator new(size_t sz, const char*, const char*);
+  void* operator new(size_t sz) { return operator new(sz, "??(??)", nullptr); }
   void operator delete(void* ptr, size_t sz);
 };
