+++ include/Kyoto/TOneStatic.hpp
@@ -6,11 +6,11 @@
 template < typename T >
 class TOneStatic {
 public:
-  void* operator new(unsigned long sz, const char*, const char*); /* {
+  void* operator new(size_t sz, const char*, const char*); /* {
    ReferenceCount()++;
    return GetAllocSpace();
  }*/
-  void* operator new(unsigned long sz) { return operator new(sz, "??(??)", nullptr); }
+  void* operator new(size_t sz) { return operator new(sz, "??(??)", nullptr); }
   void operator delete(void* ptr);
 
 private:
